# perf: parallelize hybrid search