**Before submitting, make sure you've checked the following:**
- This is broken down part of https://github.com/open-webui/open-webui/pull/11814 which was closed

- [x]  **Target branch:** Please verify that the pull request targets the `dev` branch.
- [x]  **Description:** Provide a concise description of the changes made in this pull request.
- [x]  **Changelog:** Ensure a changelog entry following the format of [Keep a Changelog](https://keepachangelog.com/) is added at the bottom of the PR description.
- [x]  **Documentation:** Have you updated relevant documentation [Open WebUI Docs](https://github.com/open-webui/docs), or other documentation sources?
- [x]  **Dependencies:** Are there any new dependencies? Have you updated the dependency versions in the documentation?
- [x]  **Testing:** Have you written and run sufficient tests for validating the changes?
- [x]  **Code review:** Have you performed a self-review of your code, addressing any coding standard issues and ensuring adherence to the project's coding standards?
- [x]  **Prefix:** To cleary categorize this pull request, prefix the pull request title, using one of the following

# Changelog Entry

### Description

- I noticed that RAG with hybrid search is slow. Currently, the hybrid search runs a nested 2 level for loop in serial to get all potential relevant parts of the document(s).

### Changed

#### Function query_doc_with_hybrid_search updates:

- Replace nested for loop with a ThreadPoolExecutor to get all collection data in parallel and only once for all queries.
- Execution of process_query over all potential documents in parallel

---

### Additional Information

- I measured the speed up based on log entries time stamps. 

## Performance Benchmark Results  

| Cores Allowed | Without PR (s) | With PR (s) | Improvement (%) |
|--------------|---------------|------------|----------------|
| 2            | 244.96        | 246.78      | -0.7% (insignificant) |
| 8            | 82.79         | 64.93 (avg) | **21.6%**      |
| 32           | 60.36         | 37.89       | **37.2%**      |

> *Note:* 8-core "With PR" time is an average of three runs which had a variance of 0.21%. The 0.7% slowdown on 2 cores is mostly within test variance.  

### Conclusion  
- **Strong speedup on 8+ cores:** +21.6% (8 cores), +37.2% (32 cores).  
- **No significant effect on 2 cores** (likely test variance).  
- **PR scales well with available cores, optimal for 8+ cores.**  
- The **default settings of the ThreadPoolExecutor** do not "spam" a low core system.