@thinkbuildlaunch 2025-04-01
hello, would like to request either an env variable to disable new hybrid search logic or a fix to the following if possible:
latest update breaks compatibility with historical pgvector data when used with IVFFLAT and reranker enabled. Previously, vectors with inconsistent lengths (e.g. ≠ 768) worked fine—either padded silently or ignored during ranking. The new hybrid search runs in parallel and includes a reranking step that now assumes all results have valid float scores and non-null text. This causes a TypeError when malformed or short vectors exist in the collection. The system fails at runtime instead of skipping bad rows like before. This regression wasn’t present before parallel hybrid + reranker changes.