@Phlogi 2025-04-01
@thinkbuildlaunch would you be able to test a PR that fixes it? instead of making this a config we should fix the root cause.