@thinkbuildlaunch 2025-04-01
I take it back, the tools I would need to attempt such a PR would be on my main setup, i think i will try a 1536 embedding first to avoid this problem and then i could try