@thinkbuildlaunch 2025-04-04
@Phlogi ok some more context..may have more to do with my use of a cross encoder reranker. see AI assisted diagnosis:

– thanks for the performance improvements in this PR. I wanted to follow up with some findings after testing the new parallel hybrid search logic on a clean setup.

Here’s what I’ve found so far:

## Environment

- Open WebUI (latest version)
- OpenAI model: `gpt-4o-latest` (default connector)
- Vector store: `pgvector` with IVFFLAT (also reproduced with `sqlite`)
- Embedding model: `BAAI/bge-large-en` (also tested with 768D variants)
- Reranker: `cross-encoder/ms-marco-MiniLM-L-12-v2`
- Hybrid search enabled (default behavior)

## Behavior

Crashes occur during hybrid search when one or more of the retrieval results are `None` or missing expected structure. This happens even when user prompts are valid, well-formed, and purely text-based.

Observed conditions:
- `collection_result = None` in some parallel threads
- `.documents` may be empty or missing
- Some retrieval hits return items without proper `.text` or `.score` fields

This results in:
```python
collection_result.documents[0]  # → AttributeError: 'NoneType' object has no attribute 'documents'
```
Because the failure occurs in a thread (inside `process_query()`), it bubbles up unhandled. On Docker systems, this can lead to container crashes or unexpected restarts.

## Additional Notes

- The issue is not tied to query quality or user input.
- It appears to arise when hybrid search operates on content that includes residual or non-standard data from prior context (e.g., after image prompts or just in normal multi-turn sessions).
- With the old serial loop, these cases were either missed or skipped silently.
- With parallel execution, they surface more readily.

## Suggested Fix

In `query_doc_with_hybrid_search()`:
```python
if not collection_result or not getattr(collection_result, "documents", None):
    return None
```

And optionally, before reranking:
```python
if not isinstance(text, str) or not text.strip():
    continue
```

## Why This Matters

- The problem can occur with both `pgvector` and `sqlite` setups.
- It is not caused by legacy data — I’ve reproduced it with fully rebuilt collections and uniform embedding dimensions.
- It mostly affects any setup using hybrid search alongside strict rerankers that expect fully structured inputs.

Let me know if you’d like crash logs or additional repro steps. I’m happy to help verify a fix once in place — and overall, the performance gain from this PR is great. This would make it much more robust under edge cases.
```