@thinkbuildlaunch 2025-04-04
I am no longer using a crossencoder reranker (for now...because i love their performance,) and I can confirm seeing tracebacks for ERROR    | open_webui.retrieval.utils:process_query:320 - Error when querying the collection with hybrid_search: 'NoneType' object has no attribute 'documents' - {}

however, it tolerates it and does not crash. 

`2025-04-04 09:48:48 2025-04-04 13:48:48.282 | INFO     | open_webui.retrieval.utils:query_collection_with_hybrid_search:302 - Starting hybrid search for 3 queries in 1 collections... - {}
2025-04-04 09:48:48 2025-04-04 13:48:48.282 | ERROR    | open_webui.retrieval.utils:process_query:320 - Error when querying the collection with hybrid_search: 'NoneType' object has no attribute 'documents' - {}
2025-04-04 09:48:48 Traceback (most recent call last):
2025-04-04 09:48:48 
2025-04-04 09:48:48   File "/usr/local/lib/python3.11/threading.py", line 1002, in _bootstrap
2025-04-04 09:48:48     self._bootstrap_inner()
2025-04-04 09:48:48     │    └ <function Thread._bootstrap_inner at 0x7f252c664860>
2025-04-04 09:48:48     └ <Thread(ThreadPoolExecutor-6_0, started 139791626651328)>
2025-04-04 09:48:48   File "/usr/local/lib/python3.11/threading.py", line 1045, in _bootstrap_inner
2025-04-04 09:48:48     self.run()
2025-04-04 09:48:48     │    └ <function Thread.run at 0x7f252c664540>
2025-04-04 09:48:48     └ <Thread(ThreadPoolExecutor-6_0, started 139791626651328)>
2025-04-04 09:48:48   File "/usr/local/lib/python3.11/threading.py", line 982, in run
2025-04-04 09:48:48     self._target(*self._args, **self._kwargs)
2025-04-04 09:48:48     │    │        │    │        │    └ {}
2025-04-04 09:48:48     │    │        │    │        └ <Thread(ThreadPoolExecutor-6_0, started 139791626651328)>
2025-04-04 09:48:48     │    │        │    └ (<weakref at 0x7f23c7bdee80; to 'ThreadPoolExecutor' at 0x7f243481af50>, <_queue.SimpleQueue object at 0x7f23c7de6a70>, None,...
2025-04-04 09:48:48     │    │        └ <Thread(ThreadPoolExecutor-6_0, started 139791626651328)>
2025-04-04 09:48:48     │    └ <function _worker at 0x7f252b7399e0>
2025-04-04 09:48:48     └ <Thread(ThreadPoolExecutor-6_0, started 139791626651328)>
2025-04-04 09:48:48   File "/usr/local/lib/python3.11/concurrent/futures/thread.py", line 83, in _worker
2025-04-04 09:48:48     work_item.run()
2025-04-04 09:48:48     │         └ <function _WorkItem.run at 0x7f252b739b20>
2025-04-04 09:48:48     └ <concurrent.futures.thread._WorkItem object at 0x7f23c7a38a90>
2025-04-04 09:48:48   File "/usr/local/lib/python3.11/concurrent/futures/thread.py", line 58, in run
2025-04-04 09:48:48     result = self.fn(*self.args, **self.kwargs)
2025-04-04 09:48:48              │    │   │    │       │    └ {}
2025-04-04 09:48:48              │    │   │    │       └ <concurrent.futures.thread._WorkItem object at 0x7f23c7a38a90>
2025-04-04 09:48:48              │    │   │    └ ('59311721-7630-4a19-b77a-7801e76e90b0', 'best non cross encoder reranker models for hybrid search')
2025-04-04 09:48:48              │    │   └ <concurrent.futures.thread._WorkItem object at 0x7f23c7a38a90>
2025-04-04 09:48:48              │    └ <function query_collection_with_hybrid_search.<locals>.process_query at 0x7f23c5b7f880>
2025-04-04 09:48:48              └ <concurrent.futures.thread._WorkItem object at 0x7f23c7a38a90>
2025-04-04 09:48:48 
2025-04-04 09:48:48 > File "/app/backend/open_webui/retrieval/utils.py", line 308, in process_query
2025-04-04 09:48:48     result = query_doc_with_hybrid_search(
2025-04-04 09:48:48              └ <function query_doc_with_hybrid_search at 0x7f2432c85b20>
2025-04-04 09:48:48 
2025-04-04 09:48:48   File "/app/backend/open_webui/retrieval/utils.py", line 171, in query_doc_with_hybrid_search
2025-04-04 09:48:48     raise e
2025-04-04 09:48:48 
2025-04-04 09:48:48   File "/app/backend/open_webui/retrieval/utils.py", line 120, in query_doc_with_hybrid_search
2025-04-04 09:48:48     texts=collection_result.documents[0],
2025-04-04 09:48:48           └ None
2025-04-04 09:48:48 `